added back front architeckture. fixed problems, added boilerplates, new methods
