+++ backend/src/api/controllers/__init__.py
@@ -1,0 +1,5 @@
+from backend.src.api.controllers.record_controller import RecordController
+from backend.src.api.controllers.report_controller import ReportController
+from backend.src.api.controllers.task_controller import TaskController
+
+__all__ = ["ReportController", "RecordController", "TaskController"]
