+++ backend/src/api/controllers/profile_controller.py
@@ -1,0 +1,5 @@
+from litestar import Controller
+
+
+class ProfileController(Controller):
+    pass
