+++ backend/src/api/dto/__init__.py
@@ -1,0 +1,29 @@
+from backend.src.api.dto.record_dto import (
+    DailyRecordRequest,
+    DailyRecordRequestDTO,
+    DailyRecordResponse,
+    DailyRecordResponseDTO,
+)
+from backend.src.api.dto.report_dto import (
+    DailyReportRequest,
+    DailyReportRequestDTO,
+    DailyReportRequestUpdate,
+    DailyReportResponse,
+    DailyReportResponseDTO,
+)
+from backend.src.api.dto.user_dto import UserReadDTO, UserRegistrationDTO, UserUpdateDTO
+
+__all__ = [
+    "DailyReportRequest",
+    "DailyReportResponse",
+    "DailyRecordRequestDTO",
+    "DailyRecordResponseDTO",
+    "DailyRecordRequest",
+    "DailyRecordResponse",
+    "DailyReportRequestDTO",
+    "DailyReportResponseDTO",
+    "UserReadDTO",
+    "UserRegistrationDTO",
+    "UserUpdateDTO",
+    "DailyReportRequestUpdate",
+]
