+++ backend/src/database/repositories/__init__.py
@@ -1,0 +1,19 @@
+from backend.src.database.repositories.external_system_repository import (
+    ExternalSystemRepository,
+)
+from backend.src.database.repositories.external_task_repository import (
+    ExternalTaskRepository,
+)
+from backend.src.database.repositories.record_repository import (
+    DailyRecordRepository,
+)
+from backend.src.database.repositories.report_repository import (
+    DailyReportRepository,
+)
+
+__all__ = [
+    "ExternalSystemRepository",
+    "ExternalTaskRepository",
+    "DailyRecordRepository",
+    "DailyReportRepository",
+]
