+++ backend/src/database/repositories/external_system_repository.py
@@ -1,0 +1,7 @@
+from advanced_alchemy import repository
+
+from backend.src.database.models import ExternalSystem
+
+
+class ExternalSystemRepository(repository.SQLAlchemyAsyncRepository[ExternalSystem]):  # type: ignore
+    model_type: type[ExternalSystem] = ExternalSystem
