+++ backend/src/integrations/ai_service.py
@@ -1,0 +1,2 @@
+class AIService:
+    """Service for working with AI API."""
