+++ backend/src/integrations/task_tracker_service.py
@@ -1,0 +1,2 @@
+class TaskTrackerService:
+    """Service for working with Task trackers systems."""
