+++ backend/src/services/__init__.py
@@ -1,0 +1,4 @@
+from backend.src.services.record_service import RecordService
+from backend.src.services.report_service import ReportService
+
+__all__ = ["RecordService", "ReportService"]
