+++ backend/src/services/profile_service.py
@@ -1,0 +1,2 @@
+class ProfileService:
+    pass
