+++ backend/src/services/task_service.py
@@ -1,0 +1,2 @@
+class TaskService:
+    pass
