--- src/api/controllers/__init__.py
@@ -1,5 +1,0 @@
-from src.api.controllers.record_controller import RecordController
-from src.api.controllers.report_controller import ReportController
-from src.api.controllers.task_controller import TaskController
-
-__all__ = ["ReportController", "RecordController", "TaskController"]
