--- src/database/repositories/external_system_repository.py
@@ -1,7 +1,0 @@
-from advanced_alchemy import repository
-
-from src.database.models import ExternalSystem
-
-
-class ExternalSystemRepository(repository.SQLAlchemyAsyncRepository[ExternalSystem]):  # type: ignore
-    model_type: type[ExternalSystem] = ExternalSystem
