--- src/services/__init__.py
@@ -1,4 +1,0 @@
-from src.services.record_service import RecordService
-from src.services.report_service import ReportService
-
-__all__ = ["RecordService", "ReportService"]
